Imprementado comentarios en recomendaciones y refactorizacion de archivos
