+++ db/initDB.js
@@ -11,7 +11,6 @@ async function main() {
     console.log('Borrando tablas existentes');
     await connection.query('DROP TABLE IF EXISTS reco');
     await connection.query('DROP TABLE IF EXISTS users');
-
     console.log('Creando tablas');
 
     await connection.query(`
@@ -39,7 +38,7 @@ async function main() {
             category VARCHAR(100) NOT NULL,
             spot VARCHAR(100) NOT NULL,
             openLine VARCHAR(100) NOT NULL,
-            text VARCHAR(300) NOT NULL,
+            text VARCHAR(3000) NOT NULL,
             image VARCHAR(100),
             votes VARCHAR(100) DEFAULT 0,
             created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
@@ -47,6 +46,18 @@ async function main() {
         );
 
        `);
+
+    // await connection.query(`
+    // CREATE TABLE comments (
+    //   cmmntId INTEGER PRIMARY KEY AUTO_INCREMENT,
+    //   userId INTEGER NOT NULL,
+    //   recoId INTEGER NOT NULL,
+    //   comment VARCHAR(300) NOT NULL,
+    //   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
+    //   FOREIGN KEY(userId) REFERENCES users(id),
+    //   FOREIGN KEY(recoId) REFERENCES reco(recoId)
+    // )
+    // `);
   } catch (error) {
     console.error(error);
   } finally {
